+++ NEWS.md
@@ -1,3 +1,8 @@
+# gtExtras 0.4.4
+- Add a `palette_col` argument to `gt_plt_bullet()` to accept a named column of palette colors as present in the data. Fixed #72
+- Fix internals of `gt_theme_pff()` to use `table.font.size` inside `tab_options()` rather than `tab_style()`. Fixed #74
+- Fix a few instances of `geom_line()` and `geom_v/hline()` that were throwing warnings for `ggplot2` v3.4.0
+
 # gtExtras 0.4.3
 - Modify internals of `gt_fa_column()` to accept factors with levels not present in the data, ignoring unneeded levels. Thank you to @mikedolanfliss for the suggestion.
 - Add `gt_merge_stack_color()` to create a merge/stack with background color - per @mrcaseb and issue #71